@@ -1,4 +1,4 @@
-// swift-tools-version: 6.0
+// swift-tools-version: 6.1
 
 import PackageDescription
 import CompilerPluginSupport
@@ -15,7 +15,7 @@ let package = Package(
     dependencies: [
         .package(name: "Core", path: "../Core"),
         .package(url: "https://github.com/realm/SwiftLint", from: "0.57.1"),
-        .package(url: "https://github.com/swiftlang/swift-syntax.git", from: "600.0.0-latest"),
+        .package(url: "https://github.com/swiftlang/swift-syntax.git", from: "601.0.0-latest"),
     ],
     targets: [
         .macro(
