@@ -1,4 +1,4 @@
-// swift-tools-version: 6.0
+// swift-tools-version: 6.1
 
 import PackageDescription
 
@@ -18,6 +18,7 @@ let package = Package(
     dependencies: [
         .package(name: "Core", path: "../Core"),
         .package(name: "Model", path: "../Model"),
+        .package(url: "https://github.com/pointfreeco/swift-dependencies", from: "1.0.0"),
         .package(url: "https://github.com/realm/SwiftLint", from: "0.57.1")
     ],
     targets: [
@@ -41,6 +42,7 @@ let package = Package(
             dependencies: [
                 "Storage",
                 .product(name: "Core", package: "Core"),
+                .product(name: "Dependencies", package: "swift-dependencies"),
                 .product(name: "Model", package: "Model")
             ],
             plugins: [.plugin(name: "SwiftLintBuildToolPlugin", package: "SwiftLint")]
