@@ -7,42 +7,57 @@
 
 import Convenience
 import Core
+import Dependencies
 import Model
 import Storage
+import StorageImplementation
 
-@Reducer public final class AppReducer {
-    public enum State: Equatable {
+@Reducer public actor AppReducer {
+    public enum State: Equatable, Sendable {
         case idle
         case loading
         case loaded([PlankRecord])
+        case error(EquatableError)
     }
     public enum Action: Sendable {
         case onAppear
-        case display([PlankRecord])
+        case display(Result<[PlankRecord], Error>)
     }
-    @Inject var storage: PlanxStorage
+    @Dependency(\.planxStorage) var storage
+}
 
-    // MARK: - Reduce Methods
+// MARK: - Reduce Methods
 
+nonisolated extension AppReducer {
     public func reduce(state: inout State, action: Action) -> Effect<Action> {
         switch action {
         case .onAppear:
             return reduceOnAppear(&state)
-        case .display(let records):
-            return reduceDisplay(&state, records: records)
+        case .display(let result):
+            return reduceDisplay(&state, result: result)
         }
     }
 
     private func reduceOnAppear(_ state: inout State) -> Effect<Action> {
         state = .loading
         return .async { [weak self] in
-            let records = await self?.storage.load()
-            return .display(records ?? [])
+            do {
+                let records = try await self?.storage.load()
+                return .display(.success(records ?? []))
+            } catch {
+                return .display(.failure(error))
+            }
         }
     }
 
-    private func reduceDisplay(_ state: inout State, records: [PlankRecord]) -> Effect<Action> {
-        state = .loaded(records)
-        return .none
+    private func reduceDisplay(_ state: inout State, result: Result<[PlankRecord], Error>) -> Effect<Action> {
+        switch result {
+        case .success(let records):
+            state = .loaded(records)
+            return .none
+        case .failure(let error):
+            state = .error(error.equatable)
+            return .none
+        }
     }
 }
