@@ -8,7 +8,8 @@
 import Foundation
 import Model
 
+// TODO: Replace this with reducers
 public protocol PlanxStorage: Sendable {
-    func load() async -> [PlankRecord]
+    func load() async throws -> [PlankRecord]
     func record(duration: TimeInterval, date: Date, feedback: Feedback?) async throws
 }
