@@ -5,9 +5,13 @@
 //  Created by Adam Londa on 16.11.2024.
 //
 
-import Assemble
 import Core
+import Dependencies
+import Model
+import ModelMocks
 import Reducers
+import Storage
+import StorageMocks
 import SwiftUI
 
 public struct AppView: View {
@@ -35,15 +39,43 @@ public struct AppView: View {
             Text("No data, yet")
         case .loaded(let data):
             Text("Loaded \(data.count) items")
+        case .error(let error):
+            Text("An error occurred: \(error.localizedDescription)")
         }
     }
 }
 
 // MARK: - Previews
 
-#Preview {
-    WithReducer(.idle,
-        dependencies: { await .mocked },
-        display: { AppView(store: $0) }
-    )
+fileprivate extension AppView {
+    init(_ planxStorage: PlanxStorage) {
+        self.init(
+            store: withDependencies {
+                $0.planxStorage = planxStorage
+            } operation: {
+                Store<AppReducer>(initialState: .idle)
+            }
+        )
+    }
+}
+
+#Preview("Empty") {
+    AppView(.loadSuccessful([]))
+}
+
+#Preview("Loaded") {
+    let now = Date.now
+    let items = [
+        PlankRecord.today(now: now, feedback: .perfect),
+        PlankRecord.yesterday(now: now, calendar: .current, feedback: .hard)
+    ]
+    AppView(.loadSuccessful(items))
+}
+
+#Preview("Loading Error") {
+    AppView(.loadFailed(.unauthorizedHealthKitAccess))
+}
+
+#Preview("Loading") {
+    AppView(.neverLoading)
 }
